Change mysql_db_insert_ret_id to return uint64_t instead of an int mysql_insert_id was always returning a my_ulonglong so we are just making it more compliant.
This is needed primarily for a following commit that will change the
job db_index to be be 64bit.

+++ src/database/mysql_common.c
@@ -896,9 +896,9 @@ extern int mysql_db_query_check_after(mysql_conn_t *mysql_conn, char *query)
 	return rc;
 }
 
-extern int mysql_db_insert_ret_id(mysql_conn_t *mysql_conn, char *query)
+extern uint64_t mysql_db_insert_ret_id(mysql_conn_t *mysql_conn, char *query)
 {
-	int new_id = 0;
+	uint64_t new_id = 0;
 
 	slurm_mutex_lock(&mysql_conn->lock);
 	if (_mysql_query_internal(mysql_conn->db_conn, query) != SLURM_ERROR)  {
